Enum vfront_rtlil::token::TokenKind

pub enum TokenKind {

    Whitespace,
    Newline,
    LineComment,
    Unknown,
    End,
    IdString,
    KwUnknown,
    Int,
    BitString,
    String,
    StringUnclosed,
    LBracket,
    RBracket,
    LBrace,
    RBrace,
    Comma,
    Colon,
    KwAutoIdx,
    KwModule,
    KwAttribute,
    KwParameter,
    KwWire,
    KwMemory,
    KwCell,
    KwProcess,
    KwConnect,
    KwSwitch,
    KwCase,
    KwAssign,
    KwSync,
    KwUpdate,
    KwMemWr,
    KwEnd,
    KwSigned,
    KwReal,
    KwWidth,
    KwUpto,
    KwOffset,
    KwSize,
    KwInput,
    KwOutput,
    KwInout,
    KwLow,
    KwHigh,
    KwPosEdge,
    KwNegEdge,
    KwEdge,
    KwAlways,
    KwGlobal,
    KwInit,

}

Type of a Token.

Variants

Whitespace

Any amount of whitespace that is not a newline

Newline

\r, \n, or \r\n

LineComment

Unknown

An unrecognized character (non-printable, non-ASCII, or loose backtick)

End

End of all input

IdString

\id or $id

KwUnknown

Looks like keyword, but isn’t.

Int

Optional sign + decimal digits

BitString

A bit string literal

String

A string literal

StringUnclosed

A string literal that is missing the closing “

LBracket

RBracket

LBrace

RBrace

Comma

Colon

KwAutoIdx

KwModule

KwAttribute

KwParameter

KwWire

KwMemory

KwCell

KwProcess

KwConnect

KwSwitch

KwCase

KwAssign

KwSync

KwUpdate

KwMemWr

KwEnd

KwSigned

KwReal

KwWidth

KwUpto

KwOffset

KwSize

KwInput

KwOutput

KwInout

KwLow

KwHigh

KwPosEdge

KwNegEdge

KwEdge

KwAlways

KwGlobal

KwInit